New Overdraft program, using Functions.
So I started to learn functions in codecademy, definitely is what this program needs right now rather then trying to import modules across files. 
Still needs some debugging. When the Main program enters the while loop it's supposed to exit when a balance of 0 is inputted by the user. However it is not currently doing so. Hopefully get more time to debug on my lunch break later today.