Fixed the while loop
So I had issues exiting the loop, but I figured out that I was not setting return of housekeeping() function equal to the balance. Fixed it. But now I think I have a logic issue with the math in the detail function. Math does not seem correct. I have to clock back into the day job so I'll have to come back to this.